Change serde tests to check serialized output
