+++ src/fp2/dusk/fp2.json
@@ -1,0 +1,4 @@
+{
+    "c0": "16e40954bea69030cc133b0597126df8d4d35ed26e4ed93346dcbdc306e2e92039a0d32ccd21176819a26cb9430335f2",
+    "c1": "0e9f824fdb1f0514e71b6536de9f0a54372b5d1141086cbceba11c132ed3373a28d7970aeaec2f185269b7b55184d64e"
+}
