+++ src/fp6/dusk/fp6.json
@@ -1,0 +1,14 @@
+{
+    "c0": {
+        "c0": "16e40954bea69030cc133b0597126df8d4d35ed26e4ed93346dcbdc306e2e92039a0d32ccd21176819a26cb9430335f2",
+        "c1": "0e9f824fdb1f0514e71b6536de9f0a54372b5d1141086cbceba11c132ed3373a28d7970aeaec2f185269b7b55184d64e"
+    },
+    "c1": {
+        "c0": "17625f46bda5caf5c111ac955059084f0a6e065bf1d2636cf1ea032c64c57792915323ef67db59aa3948d1889c59c453",
+        "c1": "15786f5024df3ec288b3617f5623038532dd239871e2e2e2f836c0aab5f8a1f6d530e4db35a0fbc704294bcfa776622f"
+    },
+    "c2": {
+        "c0": "0ce6582a5ae8d02f7411083db4a1b261dae2b461563aa8e2f544aad396398c7fa4f04a8eb09ddcd24f2eb9beca90a35b",
+        "c1": "062ec9a320aa258dab3fccd705daeb90c0975b2c0660a8aff1257c14e061ba6646bb59080638418b729fd1a877ee9465"
+    }
+}
