+++ src/fp6/dusk/mod.rs
@@ -4,7 +4,7 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use super::Fp2;
+use super::Fp6;
 
 #[cfg(feature = "serde")]
 mod serde_support {
@@ -14,36 +14,34 @@ mod serde_support {
 
     use super::*;
 
-    impl Serialize for Fp2 {
+    impl Serialize for Fp6 {
         fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
         where
             S: Serializer,
         {
-            let mut struct_ser = serializer.serialize_struct("Fp2", 2)?;
-            struct_ser.serialize_field("c0", &self.c0)?;
-            struct_ser.serialize_field("c1", &self.c1)?;
-            struct_ser.end()
+            let mut ser_struct = serializer.serialize_struct("Fp6", 3)?;
+            ser_struct.serialize_field("c0", &self.c0)?;
+            ser_struct.serialize_field("c1", &self.c1)?;
+            ser_struct.serialize_field("c2", &self.c2)?;
+            ser_struct.end()
         }
     }
 
-    impl<'de> Deserialize<'de> for Fp2 {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            struct Fp2Visitor;
+    impl<'de> Deserialize<'de> for Fp6 {
+        fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+            struct Fp6Visitor;
 
-            const FIELDS: &[&str] = &["c0", "c1"];
+            const FIELDS: &[&str] = &["c0", "c1", "c2"];
 
-            impl<'de> Visitor<'de> for Fp2Visitor {
-                type Value = Fp2;
+            impl<'de> Visitor<'de> for Fp6Visitor {
+                type Value = Fp6;
 
                 fn expecting(&self, formatter: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-                    formatter.write_str("a struct with fields c0 and c1")
+                    formatter.write_str("a struct with fields c0, c1 and c2")
                 }
 
                 fn visit_map<A: MapAccess<'de>>(self, mut map: A) -> Result<Self::Value, A::Error> {
-                    let (mut c0, mut c1) = (None, None);
+                    let (mut c0, mut c1, mut c2) = (None, None, None);
                     while let Some(key) = map.next_key()? {
                         match key {
                             "c0" => {
@@ -60,30 +58,47 @@ mod serde_support {
                                     c1 = Some(map.next_value()?);
                                 }
                             }
+                            "c2" => {
+                                if c2.is_some() {
+                                    return Err(SerdeError::duplicate_field("c2"));
+                                } else {
+                                    c2 = Some(map.next_value()?);
+                                }
+                            }
                             field => return Err(SerdeError::unknown_field(field, &FIELDS)),
                         }
                     }
-                    Ok(Fp2 {
+                    Ok(Fp6 {
                         c0: c0.ok_or_else(|| SerdeError::missing_field("c0"))?,
                         c1: c1.ok_or_else(|| SerdeError::missing_field("c1"))?,
+                        c2: c2.ok_or_else(|| SerdeError::missing_field("c2"))?,
                     })
                 }
             }
 
-            deserializer.deserialize_struct("Fp2", FIELDS, Fp2Visitor)
+            deserializer.deserialize_struct("Fp6", FIELDS, Fp6Visitor)
         }
     }
 
-    #[test]
-    fn serde_fp2() {
+    #[cfg(test)]
+    mod tests {
+        use alloc::boxed::Box;
+
         use rand::rngs::StdRng;
         use rand_core::SeedableRng;
 
-        let mut rng = StdRng::seed_from_u64(0xc0b);
-        let fp2 = Fp2::random(&mut rng);
-        let ser = serde_json::to_string(&fp2).unwrap();
-        let deser: Fp2 = serde_json::from_str(&ser).unwrap();
+        use super::*;
+        use crate::dusk::test_utils;
 
-        assert_eq!(fp2, deser);
+        #[test]
+        fn serde_fp6() -> Result<(), Box<dyn std::error::Error>> {
+            let mut rng = StdRng::seed_from_u64(0xc0b);
+            let fp6 = Fp6::random(&mut rng);
+            let ser = test_utils::assert_canonical_json(&fp6, include_str!("./fp6.json"))?;
+            let deser: Fp6 = serde_json::from_str(&ser).unwrap();
+
+            assert_eq!(fp6, deser);
+            Ok(())
+        }
     }
 }
